requirements.txt
Django==4.2.1
djangorestframework==3.12.4
psycopg2-binary==2.8.6
requests==2.26.0